more refactoring of text2sql metrics
