+++ src/unitxt/metrics.py
@@ -8,7 +8,7 @@ import uuid
 import warnings
 from abc import ABC, abstractmethod
 from collections import Counter, defaultdict
-from dataclasses import asdict, dataclass, field
+from dataclasses import asdict, field
 from dataclasses import fields as dataclasses_fields
 from enum import Enum
 from functools import lru_cache
@@ -65,6 +65,7 @@ from .operators import ArtifactFetcherMixin, Copy, FieldOperator, Set
 from .random_utils import get_seed
 from .settings_utils import get_settings
 from .stream import MultiStream, Stream
+from .text2sql_utils import SQLExecutionResult, SQLNonExecutionMetricResult
 from .type_utils import isoftype, parse_type_string, to_type_string
 from .types import ToolCall
 from .utils import deep_copy, recursive_copy, retry_connection_with_exponential_backoff
@@ -6437,26 +6438,6 @@ RISK_TYPE_TO_CLASS: Dict[RiskType, GraniteGuardianBase] = {
 }
 
 
-@dataclass
-class SQLExecutionResult:
-    execution_accuracy: int  # Whether the predicted and gold SQL results match exactly
-    non_empty_execution_accuracy: (
-        int  # Same as execution_accuracy but only if gold is non-empty
-    )
-    subset_non_empty_execution_accuracy: (
-        int  # Whether predicted is a subset of gold or vice versa, non-empty only
-    )
-    non_empty_gold_df: int  # Whether the gold SQL produced a non-empty dataframe
-    gold_sql_runtime: float  # Time taken to execute the gold SQL
-    predicted_sql_runtime: float  # Time taken to execute the predicted SQL
-    pred_to_gold_runtime_ratio: float  # Ratio of predicted runtime to gold runtime
-    gold_error: int  # Whether the gold SQL had an execution error
-    predicted_error: int  # Whether the predicted SQL had an execution error
-    gold_df_json: str  # JSON representation of the gold SQL result dataframe
-    predicted_df_json: str  # JSON representation of the predicted SQL result dataframe
-    error_message: str  # Error message from predicted execution if any
-
-
 class SQLExecutionAccuracy(InstanceMetric):
     sql_timeout: float = 60.0
     prediction_type = "Any"  # string representation is compared
@@ -6480,214 +6461,8 @@ class SQLExecutionAccuracy(InstanceMetric):
         "predicted_sql_runtime",
     ]
 
-    @staticmethod
-    def compare_dfs_ignore_colnames_ordered_rows(
-        df1: pd.DataFrame, df2: pd.DataFrame
-    ) -> bool:
-        df1.fillna(0, inplace=True)
-        df2.fillna(0, inplace=True)
-        if df1.shape != df2.shape:
-            return False
-        df1_sorted_rows = np.array([np.sort(row) for row in df1.values.astype(str)])
-        df2_sorted_rows = np.array([np.sort(row) for row in df2.values.astype(str)])
-        return np.array_equal(df1_sorted_rows, df2_sorted_rows)
-
-    @staticmethod
-    def compare_dfs_ignore_colnames_unordered_rows(
-        df1: pd.DataFrame, df2: pd.DataFrame
-    ) -> bool:
-        if df1.shape != df2.shape:
-            return False
-        df1_sorted = np.sort(np.sort(df1.values.astype(str), axis=1), axis=0)
-        df2_sorted = np.sort(np.sort(df2.values.astype(str), axis=1), axis=0)
-        return np.array_equal(df1_sorted, df2_sorted)
-
-    @staticmethod
-    def compare_dfs_ignore_colnames_subset(
-        df1: pd.DataFrame, df2: pd.DataFrame, ignore_row_order: bool = True
-    ) -> bool:
-        def row_to_multiset(row):
-            return Counter(str(x) for x in row)
-
-        def normalize_and_sort(df):
-            return pd.DataFrame(
-                sorted([sorted(map(str, row)) for row in df.values.tolist()])
-            )
-
-        if df1.empty or df2.empty or len(df1) != len(df2):
-            return False
-
-        subset_df, superset_df = (
-            (df1, df2) if df1.shape[1] <= df2.shape[1] else (df2, df1)
-        )
-
-        if ignore_row_order:
-            subset_df = normalize_and_sort(subset_df)
-            superset_df = normalize_and_sort(superset_df)
-
-        subset_rows = [row_to_multiset(row) for row in subset_df.values]
-        superset_rows = [row_to_multiset(row) for row in superset_df.values]
-
-        for r1, r2 in zip(subset_rows, superset_rows):
-            if not all(r1[k] <= r2.get(k, 0) for k in r1):
-                return False
-        return True
-
-    def _run_query(
-        self, sql: str, connector
-    ) -> Tuple[Optional[pd.DataFrame], float, str]:
-        import time
-
-        from func_timeout import func_timeout
-        from func_timeout.exceptions import FunctionTimedOut
-
-        try:
-            start = time.perf_counter()
-            result, error = func_timeout(
-                self.sql_timeout, connector.execute_query, args=(sql,)
-            )
-            duration = time.perf_counter() - start
-            if isinstance(result, dict) and "results" in result:
-                result = result["results"]
-            if error:
-                return None, duration, error
-            return pd.DataFrame(result), duration, ""
-        except FunctionTimedOut as e:
-            return None, 0.0, f"Timeout: {e}"
-        except Exception as e:
-            return None, 0.0, f"Error: {e}"
-
-    def _compare_results(
-        self, gold_df: pd.DataFrame, pred_df: pd.DataFrame, gold_sql: str
-    ) -> Tuple[int, int, int]:
-        subset_match = 0
-        non_empty_match = 0
-        if "ORDER BY" in gold_sql.upper():
-            match = int(self.compare_dfs_ignore_colnames_ordered_rows(pred_df, gold_df))
-            if not gold_df.empty and not pred_df.empty:
-                non_empty_match = match
-                if self.compare_dfs_ignore_colnames_subset(
-                    gold_df, pred_df, ignore_row_order=False
-                ):
-                    subset_match = 1
-        else:
-            match = int(
-                self.compare_dfs_ignore_colnames_unordered_rows(pred_df, gold_df)
-            )
-            if not gold_df.empty and not pred_df.empty:
-                non_empty_match = match
-                if self.compare_dfs_ignore_colnames_subset(
-                    gold_df, pred_df, ignore_row_order=True
-                ):
-                    subset_match = 1
-        return match, non_empty_match, subset_match
-
-    def get_sql_execution_results(
-        self, predicted_sql: str, gold_sql: str, connector
-    ) -> SQLExecutionResult:
-        from .sql_utils import sqlglot_optimized_equivalence
-
-        gold_df, gold_runtime, gold_error_msg = self._run_query(gold_sql, connector)
-        gold_error = int(bool(gold_error_msg))
-
-        if gold_error:
-            return SQLExecutionResult(
-                execution_accuracy=0,
-                non_empty_execution_accuracy=0,
-                subset_non_empty_execution_accuracy=0,
-                non_empty_gold_df=0,
-                gold_sql_runtime=gold_runtime,
-                predicted_sql_runtime=0,
-                pred_to_gold_runtime_ratio=0,
-                gold_error=gold_error,
-                predicted_error=0,
-                gold_df_json="",
-                predicted_df_json="",
-                error_message=gold_error_msg,
-            )
-
-        non_empty_gold_df = int(not gold_df.empty)
-        if predicted_sql.strip().lower() == gold_sql.strip().lower():
-            return SQLExecutionResult(
-                execution_accuracy=1,
-                non_empty_execution_accuracy=non_empty_gold_df,
-                subset_non_empty_execution_accuracy=non_empty_gold_df,
-                non_empty_gold_df=non_empty_gold_df,
-                gold_sql_runtime=gold_runtime,
-                predicted_sql_runtime=0,
-                pred_to_gold_runtime_ratio=0,
-                gold_error=0,
-                predicted_error=0,
-                gold_df_json=gold_df.to_json(),
-                predicted_df_json="",
-                error_message="",
-            )
-
-        try:
-            if sqlglot_optimized_equivalence(gold_sql, predicted_sql):
-                return SQLExecutionResult(
-                    execution_accuracy=1,
-                    non_empty_execution_accuracy=non_empty_gold_df,
-                    subset_non_empty_execution_accuracy=non_empty_gold_df,
-                    non_empty_gold_df=non_empty_gold_df,
-                    gold_sql_runtime=gold_runtime,
-                    predicted_sql_runtime=0,
-                    pred_to_gold_runtime_ratio=0,
-                    gold_error=0,
-                    predicted_error=0,
-                    gold_df_json=gold_df.to_json(),
-                    predicted_df_json="",
-                    error_message="",
-                )
-        except Exception as e:
-            logger.info(f"Could not check SQL equivalence: {e}")
-
-        pred_df, pred_runtime, pred_error_msg = self._run_query(
-            predicted_sql, connector
-        )
-        pred_error = 1 if pred_error_msg else 0
-
-        if pred_df is None:
-            return SQLExecutionResult(
-                execution_accuracy=0,
-                non_empty_execution_accuracy=0,
-                subset_non_empty_execution_accuracy=0,
-                non_empty_gold_df=non_empty_gold_df,
-                gold_sql_runtime=gold_runtime,
-                predicted_sql_runtime=pred_runtime,
-                pred_to_gold_runtime_ratio=(pred_runtime / gold_runtime)
-                if gold_runtime > 0
-                else 0,
-                gold_error=0,
-                predicted_error=pred_error,
-                gold_df_json=gold_df.to_json(),
-                predicted_df_json="",
-                error_message=pred_error_msg,
-            )
-
-        match, non_empty_match, subset_match = self._compare_results(
-            gold_df, pred_df, gold_sql
-        )
-
-        return SQLExecutionResult(
-            execution_accuracy=match,
-            non_empty_execution_accuracy=non_empty_match,
-            subset_non_empty_execution_accuracy=subset_match,
-            non_empty_gold_df=non_empty_gold_df,
-            gold_sql_runtime=gold_runtime,
-            predicted_sql_runtime=pred_runtime,
-            pred_to_gold_runtime_ratio=(pred_runtime / gold_runtime)
-            if gold_runtime > 0
-            else 0,
-            gold_error=0,
-            predicted_error=0,
-            gold_df_json=gold_df.to_json(),
-            predicted_df_json=pred_df.to_json(),
-            error_message=pred_error_msg,
-        )
-
     def compute(self, references: List[Any], prediction: str, task_data: Dict) -> dict:
-        from .sql_utils import get_db_connector
+        from .text2sql_utils import get_db_connector, get_sql_execution_results
 
         predicted_sql = prediction.strip()
         if not predicted_sql:
@@ -6699,8 +6474,8 @@ class SQLExecutionAccuracy(InstanceMetric):
             predicted_sql = predicted_sql[: predicted_sql.find(";") + 1]
 
         db_connector = get_db_connector(task_data["db"]["db_type"])(task_data["db"])
-        result_obj = self.get_sql_execution_results(
-            predicted_sql, references[0], db_connector
+        result_obj = get_sql_execution_results(
+            predicted_sql, references[0], db_connector, self.sql_timeout
         )
 
         result = asdict(result_obj)
@@ -6710,46 +6485,22 @@ class SQLExecutionAccuracy(InstanceMetric):
         return result
 
 
-@dataclass
-class SQLNonExecutionMetricResult:
-    sqlglot_validity: int  # Whether SQL parses with sqlglot
-    sqlparse_validity: int  # Whether SQL parses with sqlparse
-    sqlglot_equivalence: int  # Semantic equivalence using sqlglot AST
-    sqlglot_optimized_equivalence: int  # Equivalence after optimization via sqlglot
-    sqlparse_equivalence: int  # Equivalence using sqlparse AST
-    sql_exact_match: int  # Exact string match of predicted and gold SQL
-    sql_syntactic_equivalence: int  # Any of the above equivalence conditions hold
-
-
 class SQLNonExecutionAccuracy(InstanceMetric):
-    reduction_map = {
-        "mean": [
-            "sqlglot_validity",
-            "sqlparse_validity",
-            "sqlglot_equivalence",
-            "sqlglot_optimized_equivalence",
-            "sqlparse_equivalence",
-            "sql_exact_match",
-            "sql_syntactic_equivalence",
-        ]
-    }
-    main_score = "sqlglot_equivalence"
-    ci_scores = [
-        "sqlglot_validity",
-        "sqlparse_validity",
-        "sqlglot_equivalence",
-        "sqlglot_optimized_equivalence",
-        "sqlparse_equivalence",
-        "sql_exact_match",
-        "sql_syntactic_equivalence",
+    all_metrics = [
+        f.name
+        for f in dataclasses_fields(SQLNonExecutionMetricResult)
+        if isinstance(f.type, type) and f.type in (int, float)
     ]
+    reduction_map = {"mean": all_metrics}
+    main_score = "sqlglot_equivalence"
+    ci_scores = all_metrics
 
     prediction_type = "Any"  # string representation is compared
 
     _requirements_list = ["sqlglot", "sqlparse"]
 
     def compute(self, references: List[Any], prediction: str, task_data: Dict) -> dict:
-        from .sql_utils import (
+        from .text2sql_utils import (
             is_sqlglot_parsable,
             is_sqlparse_parsable,
             sql_exact_match,
